divided functions in the client
