+++ Client/Classes/AudioHandler/Receiver.cs
@@ -1,7 +1,9 @@
+using System.Net.Sockets;
+using Client.Const;
 using Client.Interfaces;
 using NAudio.Wave;
 
-namespace Client.Classes;
+namespace Client.Classes.AudioHandler;
 
 public class Receiver : IReceiver
 {
@@ -27,5 +29,25 @@ public class Receiver : IReceiver
     {
         waveOut.Stop();
         waveOut.Dispose();
+    }
+    public async Task ReceiveAudioFromServer(NetworkStream stream, IReceiver receiver)
+    {
+        byte[] buffer = new byte[4096];
+        while (true)
+        {
+            try
+            {
+                int bytesRead = await stream.ReadAsync(buffer, 0, buffer.Length);
+                if (bytesRead > 0)
+                {
+                    receiver.PlayAudio(buffer, 0, bytesRead);
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"{Constants.ErrorMessage} {ex.Message}");
+                break;
+            }
+        }
     }
 }
