+++ Client/Interfaces/ISender.cs
@@ -1,3 +1,5 @@
+using System.Net.Sockets;
+
 namespace Client.Interfaces;
 
 public interface ISender
@@ -6,4 +8,6 @@ public interface ISender
     void Stop();
     int ReadAudio(byte[] outputBuffer, int offset, int count);
     bool IsDataAvailable();
+    Task SendFullAudioToServer(NetworkStream stream, IFullAudioMaker fullAudioMaker);
+    Task TransmitAudioToServer(NetworkStream stream, ISender sender);
 }
