+++ server/Classes/ClientManager.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Sockets;
 using System.Threading.Tasks;
+using server.Const;
 using server.Interface;
 
 namespace server.Classes;
