+++ server/Interface/IGridFsManager.cs
@@ -3,4 +3,5 @@ namespace server.Interface;
 public interface IGridFsManager
 {
     Task SaveAudioAsync(string filename, byte[] audioData);
+    Task<byte[]> GetAudioAsync(string filename);
 }
