+++ server/Program.cs
@@ -1,7 +1,10 @@
 ﻿using server.Classes;
+using server.Classes.AudioHandler;
+using server.Const;
 
+var baseDirectory = AppDomain.CurrentDomain.BaseDirectory;
 var clientManager = new ClientManager();
-var gridFsManager = new GridFsManager(Constants.MongoConnectionString, Constants.DatabaseName);
+var gridFsManager = new GridFsManager(baseDirectory);
 var transmitAudio = new TransmitAudio(clientManager);
 var receiveAudio = new ReceiveAudio(transmitAudio, gridFsManager);
 var server = new TcpServer(Constants.ServerPort, clientManager, receiveAudio);
